Move spec out of aggregator
