+++ server/sink/src/test/java/org/bithon/server/sink/metrics/MetricAggregatorTest.java
@@ -19,12 +19,12 @@ package org.bithon.server.sink.metrics;
 import org.bithon.server.commons.time.TimeSpan;
 import org.bithon.server.storage.datasource.DataSourceSchema;
 import org.bithon.server.storage.datasource.TimestampSpec;
-import org.bithon.server.storage.datasource.aggregator.spec.LongMaxMetricSpec;
-import org.bithon.server.storage.datasource.aggregator.spec.LongMinMetricSpec;
-import org.bithon.server.storage.datasource.aggregator.spec.LongSumMetricSpec;
 import org.bithon.server.storage.datasource.dimension.StringDimensionSpec;
 import org.bithon.server.storage.datasource.input.IInputRow;
 import org.bithon.server.storage.datasource.input.InputRow;
+import org.bithon.server.storage.datasource.spec.max.LongMaxMetricSpec;
+import org.bithon.server.storage.datasource.spec.min.LongMinMetricSpec;
+import org.bithon.server.storage.datasource.spec.sum.LongSumMetricSpec;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
