--- server/storage/src/main/java/org/bithon/server/storage/datasource/aggregator/spec/CountMetricSpec.java
@@ -1,152 +1,0 @@
-/*
- *    Copyright 2020 bithon.org
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.bithon.server.storage.datasource.aggregator.spec;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import org.bithon.server.storage.datasource.DataSourceSchema;
-import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
-import org.bithon.server.storage.datasource.api.IQueryStageAggregator;
-import org.bithon.server.storage.datasource.api.QueryStageAggregators;
-import org.bithon.server.storage.datasource.typing.IValueType;
-import org.bithon.server.storage.datasource.typing.LongValueType;
-
-import javax.annotation.Nullable;
-import javax.validation.constraints.NotNull;
-
-/**
- * @author frank.chen021@outlook.com
- * @date 2020/12/23
- */
-public class CountMetricSpec implements IMetricSpec {
-
-    public static final IMetricSpec INSTANCE = new CountMetricSpec("count", "count");
-
-    @Getter
-    private final String name;
-
-    @Getter
-    private final String field;
-    private final IQueryStageAggregator queryStageAggregator;
-
-    @JsonCreator
-    public CountMetricSpec(@JsonProperty("name") @NotNull String name,
-                           @JsonProperty("field") @Nullable String field) {
-        this.name = name;
-        this.field = field;
-        this.queryStageAggregator = new QueryStageAggregators.CountAggregator(name, field);
-    }
-
-    @JsonIgnore
-    @Override
-    public String getType() {
-        return IMetricSpec.COUNT;
-    }
-
-    @Override
-    public boolean isVisible() {
-        return false;
-    }
-
-    @Override
-    public String getDisplayText() {
-        return "次数";
-    }
-
-    @Override
-    public String getUnit() {
-        return "次";
-    }
-
-    @Override
-    public IValueType getValueType() {
-        return LongValueType.INSTANCE;
-    }
-
-    @Override
-    public void setOwner(DataSourceSchema dataSource) {
-    }
-
-    @Override
-    public String validate(Object input) {
-        return null;
-    }
-
-    @Override
-    public <T> T accept(IMetricSpecVisitor<T> visitor) {
-        return visitor.visit(this);
-    }
-
-    @Override
-    public NumberAggregator createAggregator() {
-        return new NumberAggregator() {
-            private long value;
-
-            @Override
-            public int intValue() {
-                return (int) value;
-            }
-
-            @Override
-            public long longValue() {
-                return value;
-            }
-
-            @Override
-            public float floatValue() {
-                return value;
-            }
-
-            @Override
-            public double doubleValue() {
-                return value;
-            }
-
-            @Override
-            public void aggregate(long timestamp, Object value) {
-                this.value++;
-            }
-
-            @Override
-            public Number getNumber() {
-                return value;
-            }
-        };
-    }
-
-    @JsonIgnore
-    @Override
-    public IQueryStageAggregator getQueryAggregator() {
-        return queryStageAggregator;
-    }
-
-    @Override
-    public int hashCode() {
-        return name.hashCode();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj instanceof CountMetricSpec) {
-            return this.name.equals(((CountMetricSpec) obj).name);
-        } else {
-            return false;
-        }
-    }
-}
