+++ server/storage/src/main/java/org/bithon/server/storage/datasource/spec/max/LongMaxMetricSpec.java
@@ -14,17 +14,14 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.aggregator.spec;
+package org.bithon.server.storage.datasource.spec.max;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import org.bithon.server.storage.datasource.DataSourceSchema;
 import org.bithon.server.storage.datasource.aggregator.LongMaxAggregator;
 import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
-import org.bithon.server.storage.datasource.api.IQueryStageAggregator;
-import org.bithon.server.storage.datasource.api.QueryStageAggregators;
+import org.bithon.server.storage.datasource.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.typing.IValueType;
 import org.bithon.server.storage.datasource.typing.LongValueType;
 
@@ -35,23 +32,7 @@ import javax.validation.constraints.NotNull;
  * @author frank.chen021@outlook.com
  * @date 2021/3/16
  */
-public class LongMaxMetricSpec implements IMetricSpec {
-
-    @Getter
-    private final String name;
-
-    @Getter
-    private final String field;
-
-    @Getter
-    private final String displayText;
-
-    @Getter
-    private final String unit;
-
-    @Getter
-    private final boolean visible;
-    private final IQueryStageAggregator queryStageAggregator;
+public class LongMaxMetricSpec extends MaxMetricSpec {
 
     @JsonCreator
     public LongMaxMetricSpec(@JsonProperty("name") @NotNull String name,
@@ -59,12 +40,7 @@ public class LongMaxMetricSpec implements IMetricSpec {
                              @JsonProperty("displayText") @NotNull String displayText,
                              @JsonProperty("unit") @NotNull String unit,
                              @JsonProperty("visible") @Nullable Boolean visible) {
-        this.name = name;
-        this.field = field;
-        this.displayText = displayText;
-        this.unit = unit;
-        this.visible = visible == null ? true : visible;
-        this.queryStageAggregator = new QueryStageAggregators.MaxAggregator(name, field);
+        super(name, field, displayText, unit, visible);
     }
 
     @JsonIgnore
@@ -79,33 +55,8 @@ public class LongMaxMetricSpec implements IMetricSpec {
     }
 
     @Override
-    public void setOwner(DataSourceSchema dataSource) {
-    }
-
-    @Override
-    public String validate(Object input) {
-        return null;
-    }
-
-    @Override
-    public <T> T accept(IMetricSpecVisitor<T> visitor) {
-        return visitor.visit(this);
-    }
-
-    @Override
     public NumberAggregator createAggregator() {
         return new LongMaxAggregator();
-    }
-
-    @JsonIgnore
-    @Override
-    public IQueryStageAggregator getQueryAggregator() {
-        return queryStageAggregator;
-    }
-
-    @Override
-    public int hashCode() {
-        return name.hashCode();
     }
 
     @Override
