+++ server/storage/src/main/java/org/bithon/server/storage/datasource/spec/min/LongMinMetricSpec.java
@@ -14,17 +14,14 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.aggregator.spec;
+package org.bithon.server.storage.datasource.spec.min;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import org.bithon.server.storage.datasource.DataSourceSchema;
-import org.bithon.server.storage.datasource.aggregator.LongSumAggregator;
+import org.bithon.server.storage.datasource.aggregator.LongMinAggregator;
 import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
-import org.bithon.server.storage.datasource.api.IQueryStageAggregator;
-import org.bithon.server.storage.datasource.api.QueryStageAggregators;
+import org.bithon.server.storage.datasource.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.typing.IValueType;
 import org.bithon.server.storage.datasource.typing.LongValueType;
 
@@ -33,44 +30,23 @@ import javax.validation.constraints.NotNull;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2020/11/30 5:38 下午
+ * @date 2021/3/16
  */
-public class LongSumMetricSpec implements IMetricSpec {
-
-    @Getter
-    private final String name;
-
-    @Getter
-    private final String field;
-
-    @Getter
-    private final String displayText;
-
-    @Getter
-    private final String unit;
-
-    @Getter
-    private final boolean visible;
-    private final IQueryStageAggregator queryStageAggregator;
+public class LongMinMetricSpec extends MinMetricSpec {
 
     @JsonCreator
-    public LongSumMetricSpec(@JsonProperty("name") @NotNull String name,
+    public LongMinMetricSpec(@JsonProperty("name") @NotNull String name,
                              @JsonProperty("field") @Nullable String field,
                              @JsonProperty("displayText") @NotNull String displayText,
                              @JsonProperty("unit") @NotNull String unit,
                              @JsonProperty("visible") @Nullable Boolean visible) {
-        this.name = name;
-        this.field = field;
-        this.displayText = displayText;
-        this.unit = unit;
-        this.visible = visible == null ? true : visible;
-        this.queryStageAggregator = new QueryStageAggregators.SumAggregator(name, field);
+        super(name, field, displayText, unit, visible);
     }
 
     @JsonIgnore
     @Override
     public String getType() {
-        return LONG_SUM;
+        return IMetricSpec.LONG_MIN;
     }
 
     @Override
@@ -79,39 +55,14 @@ public class LongSumMetricSpec implements IMetricSpec {
     }
 
     @Override
-    public void setOwner(DataSourceSchema dataSource) {
-    }
-
-    @Override
-    public String validate(Object input) {
-        return null;
-    }
-
-    @Override
-    public <T> T accept(IMetricSpecVisitor<T> visitor) {
-        return visitor.visit(this);
-    }
-
-    @Override
     public NumberAggregator createAggregator() {
-        return new LongSumAggregator();
-    }
-
-    @JsonIgnore
-    @Override
-    public IQueryStageAggregator getQueryAggregator() {
-        return queryStageAggregator;
-    }
-
-    @Override
-    public int hashCode() {
-        return name.hashCode();
+        return new LongMinAggregator();
     }
 
     @Override
     public boolean equals(Object obj) {
-        if (obj instanceof LongSumMetricSpec) {
-            return this.name.equals(((LongSumMetricSpec) obj).name);
+        if (obj instanceof LongMinMetricSpec) {
+            return this.name.equals(((LongMinMetricSpec) obj).name);
         } else {
             return false;
         }
