+++ server/web-service/src/main/java/org/bithon/server/web/service/api/DataSourceService.java
@@ -19,8 +19,8 @@ package org.bithon.server.web.service.api;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import org.bithon.server.commons.time.TimeSpan;
-import org.bithon.server.storage.datasource.aggregator.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.api.IQueryStageAggregator;
+import org.bithon.server.storage.datasource.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.typing.DoubleValueType;
 import org.bithon.server.storage.metrics.IMetricStorage;
 import org.bithon.server.storage.metrics.TimeseriesQuery;
